feat(site): migrate experience site to Astro
Replace the Vite vanilla site with Astro 7. Use build-time Astro Code
blocks for highlighting and keep wall.js client init + hash sync in
src/scripts. Update docs and README.
