+++ apps/site/package.json
@@ -4,18 +4,18 @@
   "version": "2.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "typecheck": "tsc --noEmit",
-    "clean": "rm -rf site-dist"
+    "dev": "astro dev --port 5173",
+    "build": "astro build",
+    "preview": "astro preview --port 4173",
+    "typecheck": "tsc --noEmit -p tsconfig.scripts.json",
+    "clean": "rm -rf site-dist .astro"
   },
   "dependencies": {
-    "shiki": "^4.3.1",
     "wall.js": "workspace:*"
   },
   "devDependencies": {
-    "typescript": "^7.0.2",
-    "vite": "^8.1.4"
+    "astro": "^7.0.7",
+    "typescript": "^7.0.2"
   }
 }
+
