+++ apps/site/src/scripts/hash-sync.ts
