+++ docs/SITE.md
@@ -1,6 +1,6 @@
 # Site (experience demo)
 
-The marketing site under `apps/site` **is itself a wall.js instance**. Scrolling the homepage exercises the core library. It is **vanilla** (Vite + TS) — not the React port.
+The marketing site under `apps/site` **is itself a wall.js instance**, built with **Astro**. Scrolling the homepage exercises the core library.
 
 ## Develop
 
@@ -9,10 +9,10 @@ From the monorepo root:
 ```bash
 pnpm install
 pnpm dev
-# → http://localhost:5173  (turbo → @wall.js/site)
+# → http://localhost:5173  (turbo → @wall.js/site → astro dev)
 ```
 
-Dev resolves `wall.js` to `packages/wall/src` for HMR.
+Dev resolves `wall.js` to `packages/wall/src` for HMR (see `astro.config.mjs`).
 
 ## Build
 
@@ -21,25 +21,28 @@ pnpm build:site
 # → apps/site/site-dist/
 ```
 
-Turbo builds `packages/wall` first (`dependsOn: ["^build"]`), then the site.
+Turbo builds `packages/wall` first (`dependsOn: ["^build"]`), then Astro.
 
 ## Structure
 
 ```
 apps/site/
-  index.html
-  styles.css
-  main.ts          # new Wall('#wall') + chrome
-  hash-sync.ts     # app-layer hash ↔ section (not core)
-  vite.config.ts
+  astro.config.mjs
+  public/                 # favicon, logos
+  src/
+    pages/index.astro     # fullpage markup + Astro <Code> blocks
+    scripts/client.ts     # new Wall('#wall') + chrome
+    scripts/hash-sync.ts  # app-layer hash ↔ section
+    styles/global.css
 ```
 
 ## Interactive bits
 
-- Section `id`s + **hash sync** (`hash-sync.ts`) — scroll updates `#features`; open `/#api` deep-links
-- Top nav `<a href="#…">` intercepted by `bindHashLinks`
-- `[data-wall-section-nav]` — right-side dots
-- Slide section with `data-wall-slide` + arrow controls
-- Progress bar driven by `wall.on('change')`
+- Section `id`s + **hash sync** (`hash-sync.ts`)
+- Top nav `<a href="#…">` via `bindHashLinks`
+- `[data-wall-section-nav]` dots
+- Horizontal `data-wall-slide` section
+- Progress bar + active nav from `wall.on('change')`
+- Syntax highlighting via Astro `<Code>` (build-time Shiki)
 
-Hash binding is intentionally **not** in wall.js core — see [API recipes](./API.md#recipes-app-layer--not-core).
+Hash binding is **not** in wall.js core — see [API recipes](./API.md#recipes-app-layer--not-core).
